+++ core/src/main/java/com/facebook/ktfmt/format/Formatter.kt
@@ -102,8 +102,9 @@ object Formatter {
         .let { convertLineSeparators(it) }
         .let { sortedAndDistinctImports(it) }
         .let { dropRedundantElements(it, options) }
-        .let { prettyPrint(it, options, "\n") }
+        .let { prettyPrint(it, options, lineSeparator = "\n") }
         .let { addRedundantElements(it, options) }
+        .let { MultilineStringFormatter(options.continuationIndent).format(it) }
         .let { convertLineSeparators(it, checkNotNull(Newlines.guessLineSeparator(kotlinCode))) }
         .let { if (shebang.isEmpty()) it else shebang + "\n" + it }
   }